LPS-197299 Move loadClientExtensions to its own module
As per https://github.com/liferay-frontend/liferay-portal/pull/3703#discussion_r1358332202
